Removing encoding/json, log, github.com/nats-io/nuid
