+++ lib/gopherjs/thrift/debug_protocol.go
@@ -1,3 +1,5 @@
+// +build !js
+
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements. See the NOTICE file
